temp fix for trending tags bug on ecosystem page (#137)
## Overview
<!--- Description of the scope of the PR. Add details what has changed
(and why if needed) --->
This PR is to quickly fix clicking trending tags break the frontend. The
issue comes from we defined projects' tags manually and some project's
tags are not in the list. Thus it has `Cannot read properties of
undefined (reading 'toLowerCase')`
We have a ticket to pull tags from airtable rather than defined tags
ourselves
## Tickets
<!--- Paste Asana/Linear tickets here --->

## Notes for reviewer (optional)
<!--- 
- (For example): I wasn't sure whether to implement a caching strategy
[here](https://…), so I left a comment and suggest that this is handled
in a separate [ticket](https://…)
--->